+++ NEWS.md
@@ -1,6 +1,8 @@
 ## Version 2.0.0.9000
 * If one goes to `help.start()` and then clicks on the "Bullock" entry, 
   "Bullock-package" is now the first entry in the help index.
+* Small improvements to documentation formatting.
+
 
 # Version 2.0.0 (2020-01-08)
 * Added `regTable()`, `latexTable()`, and `latexTablePDF()`.