+++ NEWS.md
@@ -1,5 +1,6 @@
 ## Version 2.0.0.9000
 * In `regTable()`, improved checking of classes of `objList` elements.
+* Better `latexTable()` warning message when `colNames[[1]]` has the wrong length.
 * If one goes to `help.start()` and then clicks on the "Bullock" entry, 
   "Bullock-package" is now the first entry in the help index.
 * Small improvements to documentation formatting.