+++ NEWS.md
@@ -1,5 +1,6 @@
 ## Version 2.0.0.9000
-* Fixed bug in latexTable() calls for which `decimalPlaces` is 0.
+* Fixed bug in `latexTable()` calls for which `decimalPlaces` is 0.
+* Fixed bug in float placement of landscaped tables made by `latexTable()`. 
 * In `regTable()`, improved checking of classes of `objList` elements.
 * Better `latexTable()` warning message when `colNames[[1]]` has the wrong length.
 * If one goes to `help.start()` and then clicks on the "Bullock" entry,