+++ NEWS.md
@@ -4,8 +4,9 @@
   to `-.025in`. Users can now arbitrarily adjust the vertical space between
   column-name rows by specifying any LaTeX length as the 
   `spaceBetweenColNameRows` argument.   
-* Added c.latexTable method (#12).
-* Adjusted print.regTable method so that row names are right-aligned (#11).
+* Added `c.latexTable()` method (#12).
+* Adjusted `print.regTable()` method so that row names are right-aligned (#11).
+* Reduced `lNAv()` output from two lines to 1 (#16).
 * Fixed bug in `latexTable()` calls for which `decimalPlaces` is 0 (#8).
 * Fixed bug in `latexTable()` calls that arose when `SE_table` was `FALSE` and
   an element of the `colNames` list had a length less than `ncol(mat)`.